add options

+++ src/os/bluestore/BlueStore.cc
@@ -15446,14 +15446,14 @@ void BlueStore::BlueStoreCoDel::init(CephContext* cct) {
     if (cct->_conf.get_val<bool>("bluestore_codel")) {
         activated = cct->_conf.get_val<bool>("bluestore_codel");
     }
-    if (cct->_conf.get_val<int>("bluestore_codel_target_latency")) {
-        initial_target_latency = cct->_conf.get_val<int>("bluestore_codel_target_latency");
+    if (cct->_conf.get_val<int64_t>("bluestore_codel_target_latency")) {
+        initial_target_latency = cct->_conf.get_val<int64_t>("bluestore_codel_target_latency");
     }
-    if (cct->_conf.get_val<int>("bluestore_codel_interval")) {
-        initial_interval = cct->_conf.get_val<int>("bluestore_codel_interval");
+    if (cct->_conf.get_val<int64_t>("bluestore_codel_interval")) {
+        initial_interval = cct->_conf.get_val<int64_t>("bluestore_codel_interval");
     }
-    if (cct->_conf.get_val<int>("bluestore_codel_init_batch_size")) {
-        initial_batch_size = cct->_conf.get_val<int>("bluestore_codel_init_batch_size");
+    if (cct->_conf.get_val<int64_t>("bluestore_codel_init_batch_size")) {
+        initial_batch_size = cct->_conf.get_val<int64_t>("bluestore_codel_init_batch_size");
         batch_size = initial_batch_size;
     }
     if (cct->_conf.get_val<float>("bluestore_codel_batch_size_limit_ratio")) {
